refactor: criação de pastas "topic" & "user"
