+++ lib/models/topic/topic.dart
@@ -1,4 +1,4 @@
-import 'item.dart';
+import '../item.dart';
 
 class Topic {
   final String name;
