+++ test/models/user_test.dart
@@ -1,5 +1,5 @@
 import 'package:flutter_test/flutter_test.dart';
-import 'package:soft_list/models/user.dart';
+import 'package:soft_list/models/user/user.dart';
 
 void main() {
   group('User Unit Tests', () {
